refactor(examples): separate UI from rendering
